Change the order of the players from a,b,c,S to S,a,b,c.
The previous ordering, a,b,c,S, was unnecessarily difficult because
the additional constraint (additional in the sense of additional over
and above "0 <= p_A, p_B, p_C, p_S<1" and "p_A + p_B + p_C <= 1") was
on the fillup value p_S, so "p_S<=small" was forced to appear as
"p_A+p_B+p_C>=1-small", which was confusing as the constraint on p_S
was written in terms of the other strengths.

Now we have S,a,b,c this is easier.  We have the additional constraint
appearing as the much simpler form "p_S<small".

Actually, now that I write it like this, I am not at all sure that it
is easier.  If there is an "odd" strength, maybe it is logical to make
fillup value be this odd strength.  The previous version has
p_A,p_B,p_C all on the same footing; and this commit breaks the
symmetry between A,B,C by making "C" special.  Surely if any strength
is special, it should be the fillup value?